+++ extensions/amp-a4a/0.1/test/test-amp-a4a.js
@@ -36,8 +36,8 @@ import {resetScheduledElementForTesting} from '#service/custom-element-registry'
 import {Extensions} from '#service/extensions-impl';
 import {installRealTimeConfigServiceForDoc} from '#service/real-time-config/real-time-config-impl';
 
+import {macroTask} from '#testing/helpers';
 import {createIframePromise} from '#testing/iframe';
-import {macroTask} from '#testing/yield';
 
 import {FetchMock, networkFailure} from './fetch-mock';
 import {data as testFragments} from './testdata/test_fragments';
