+++ extensions/amp-ad/0.1/test/test-amp-ad-ui.js
@@ -19,7 +19,7 @@ import * as domQuery from '#core/dom/query';
 import {AmpAdUIHandler} from '../amp-ad-ui';
 import {BaseElement} from '../../../../src/base-element';
 import {createElementWithAttributes} from '#core/dom';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 import {setStyles} from '#core/dom/style';
 
 describes.realWin(
