+++ extensions/amp-ad/0.1/test/test-concurrent-load.js
@@ -23,7 +23,7 @@ import {
   waitFor3pThrottle,
 } from '../concurrent-load';
 import {installTimerService} from '#service/timer-impl';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('concurrent-load', {}, (env) => {
   describe('getAmpAdRenderOutsideViewport', () => {
