+++ extensions/amp-analytics/0.1/test/test-browser-events.js
@@ -20,7 +20,7 @@ import {
   AnalyticsEventType,
   BrowserEventTracker,
 } from '../events';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 import {toggleExperiment} from '#experiments';
 
 describes.realWin(
