+++ extensions/amp-analytics/0.1/test/test-events.js
@@ -31,7 +31,7 @@ import {
 import {AmpdocAnalyticsRoot} from '../analytics-root';
 import {Deferred} from '#core/data-structures/promise';
 import {Signals} from '#core/data-structures/signals';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('Events', {amp: 1}, (env) => {
   let win;
