+++ extensions/amp-analytics/0.1/test/test-requests.js
@@ -22,7 +22,7 @@ import {RequestHandler, expandPostMessage} from '../requests';
 import {Services} from '#service';
 import {installLinkerReaderService} from '../linker-reader';
 import {installSessionServiceForTesting} from '../session-manager';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('Requests', {amp: 1}, (env) => {
   let ampdoc;
