+++ extensions/amp-analytics/0.1/test/test-video-analytics.js
@@ -17,7 +17,7 @@
 import {AmpdocAnalyticsRoot} from '../analytics-root';
 import {AnalyticsEvent, AnalyticsEventType, VideoEventTracker} from '../events';
 import {dispatchCustomEvent} from 'src/core/dom/index';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin(
   'Events',
