+++ extensions/amp-brightcove/0.1/test/test-amp-brightcove.js
@@ -22,7 +22,7 @@ import {CommonSignals} from '#core/constants/common-signals';
 import {VideoEvents} from '../../../../src/video-interface';
 import {createElementWithAttributes} from '#core/dom';
 import {listenOncePromise} from '../../../../src/event-helper';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 import {parseUrlDeprecated} from '../../../../src/url';
 import {whenUpgradedToCustomElement} from '#core/dom/amp-element-helpers';
