+++ extensions/amp-consent/0.1/test/test-consent-policy-manager.js
@@ -31,7 +31,7 @@ import {
 } from '../consent-policy-manager';
 import {dict} from '#core/types/object';
 import {expandPolicyConfig} from '../consent-config';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 import {
   registerServiceBuilder,
