+++ extensions/amp-consent/0.1/test/test-consent-state-manager.js
@@ -24,7 +24,7 @@ import {CONSENT_STRING_TYPE} from '#core/constants/consent-state';
 import {ConsentInstance, ConsentStateManager} from '../consent-state-manager';
 import {Services} from '#service';
 import {dev} from '../../../../src/log';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 import {
   registerServiceBuilder,
   resetServiceForTesting,
