+++ extensions/amp-next-page/0.1/test/test-amp-next-page.js
@@ -21,7 +21,7 @@ import {toggleExperiment} from '#experiments';
 
 import {Services} from '#service';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 import * as DocFetcher from '../../../../src/document-fetcher';
 import {getServicePromiseForDoc} from '../../../../src/service-helpers';
