+++ extensions/amp-sticky-ad/1.0/test/test-amp-sticky-ad.js
@@ -18,10 +18,8 @@ import '../../../amp-ad/0.1/amp-ad';
 import '../amp-sticky-ad';
 import {createElementWithAttributes} from '#core/dom';
 
-import {Services} from '#service';
-
+import {macroTask} from '#testing/helpers';
 import {poll} from '#testing/iframe';
-import {macroTask} from '#testing/yield';
 
 describes.realWin(
   'amp-sticky-ad 1.0 version',
