+++ extensions/amp-story-auto-ads/0.1/test/test-algorithm-count-pages.js
@@ -21,7 +21,7 @@ import {
 import {AmpStory} from '../../../amp-story/1.0/amp-story';
 import {CountPagesAlgorithm} from '../algorithm-count-pages';
 import {InsertionState, StoryAdPageManager} from '../story-ad-page-manager';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('CountPagesAlgorithm', {amp: true}, (env) => {
   let storeService;
