+++ extensions/amp-story-auto-ads/0.1/test/test-story-ad-page-manager.js
@@ -19,7 +19,7 @@ import {AmpStory} from '../../../amp-story/1.0/amp-story';
 import {InsertionState, StoryAdPageManager} from '../story-ad-page-manager';
 import {StoryAdPage} from '../story-ad-page';
 import {getServiceForDoc} from '../../../../src/service-helpers';
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 // TODO(ccordry): add tests for analytics events.
 describes.realWin('StoryAdPageManager', {amp: true}, (env) => {
