+++ extensions/amp-user-notification/0.1/test/test-amp-user-notification.js
@@ -16,7 +16,7 @@
 
 import {Services} from '#service';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 import {
   registerServiceBuilder,
