+++ extensions/amp-video-iframe/0.1/test/test-amp-video-iframe.js
@@ -20,11 +20,10 @@ import {whenUpgradedToCustomElement} from '#core/dom/amp-element-helpers';
 
 import {Services} from '#service';
 
+import {macroTask} from '#testing/helpers';
 import {installResizeObserverStub} from '#testing/resize-observer-stub';
-import {macroTask} from '#testing/yield';
 
 import {listenOncePromise} from '../../../../src/event-helper';
-import {VideoEvents} from '../../../../src/video-interface';
 
 describes.realWin(
   'amp-video-iframe',
