+++ test/unit/core/dom/stream/test-response.js
@@ -16,7 +16,7 @@
 
 import {streamResponseToWriter} from '#core/dom/stream';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 const chunk1 = `
  <!doctype html>
