+++ test/unit/test-impression.js
@@ -22,7 +22,7 @@ import {toggleExperiment} from '#experiments';
 
 import {Services} from '#service';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 import {
   getExtraParamsUrl,
