+++ test/unit/test-navigation.js
@@ -20,7 +20,7 @@ import {Services} from '#service';
 import {maybeExpandUrlParamsForTesting} from '#service/navigation';
 import {installUrlReplacementsServiceForDoc} from '#service/url-replacements-impl';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 import * as Impression from '../../src/impression';
 import {addParamToUrl} from '../../src/url';
