+++ test/unit/test-notification-ui-manager.js
@@ -15,7 +15,7 @@
  */
 import {NotificationUiManager} from '#service/notification-ui-manager';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('NotificationUiManager', {amp: 1}, (env) => {
   describe('Notification UI Manager', () => {
