+++ test/unit/test-yield.js
@@ -15,7 +15,7 @@
  */
 import * as fakeTimers from '@sinonjs/fake-timers';
 
-import {macroTask} from '#testing/yield';
+import {macroTask} from '#testing/helpers';
 
 describes.realWin('yield', {}, (env) => {
   let win;
